# app/api/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "Fabric CLI http wrapper"
  },
  "host": "localhost",
  "basePath": "/api",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "definitions": {
...
    "TxCount": {
      "type": "object",
      "properties": {
        "datetime": {
          "type": "string"
        },
        "count": {
          "type": "string"
        }
      }
    },
    "Error": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      }
    }
  },
  "paths": {
...
    "/status/{channel_genesis_hash}": {
      "x-swagger-router-controller": "dbroutes",
      "get": {
        "tags": ["Channel"],
        "operationId": "channelStatus",
        "summary": "Query to get BlockCount on a channel",
        "description": "Query to get BlockCount on a channel",
        "parameters": [
          {
            "name": "channel_genesis_hash",
            "in": "path",
            "description": "channel_genesis_hash",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "chaincodeCount": {
                  "type": "string"
                },
                "txCount": {
                  "type": "string"
                },
                "peerCount": {
                  "type": "string"
                },
                "latestBLock": {
                  "type": "number"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/block/transactions/{channel_genesis_hash}/{number}": {
      "x-swagger-router-controller": "dbroutes",
      "get": {
        "tags": ["Transactions"],
        "operationId": "transactionsCount",
        "summary": "Query to Fetch Block Transactions",
        "description": "Query to Fetch Block Transactions",
        "parameters": [
          {
            "name": "channel_genesis_hash",
            "in": "path",
            "description": "channel_genesis_hash",
            "required": true,
            "type": "string"
          },
          {
            "name": "number",
            "in": "path",
            "description": "block number",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number"
                },
                "number": {
                  "type": "number"
                },
                "txCount": {
                  "type": "number"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/blockActivity/{channel_genesis_hash}": {
      "x-swagger-router-controller": "dbroutes",
      "get": {
        "tags": ["Blocks"],
        "operationId": "blockActivity",
        "summary": "Query to Fetch Block Activity List",
        "description": "Query to Fetch Block Acivity List",
        "parameters": [
          {
            "name": "channel_genesis_hash",
            "in": "path",
            "description": "channel_genesis_hash",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number"
                },
                "txcount": {
                  "type": "number"
                },
                "datahash": {
                  "type": "string"
                },
                "blockhash": {
                  "type": "string"
                },
                "prehash": {
                  "type": "string"
                },
                "createdt": {
                  "type": "string"
                },
                "txhash": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/peers/{channel_genesis_hash}": {
      "x-swagger-router-controller": "dbroutes",
      "get": {
        "tags": ["Channel"],
        "operationId": "peerList",
        "summary": "Peers",
        "description": "Peers",
        "parameters": [
          {
            "name": "channel_genesis_hash",
            "in": "path",
            "description": "channel_genesis_hash",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number"
                },
                "peers": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "requests": {
                        "type": "string"
                      },
                      "channel_genesis_hash": {
                        "type": "string"
                      },
                      "server_hostname": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/peersStatus/{channelName}": {
      "x-swagger-router-controller": "platformroutes",
      "get": {
        "tags": ["Channel"],
        "operationId": "peerList",
        "summary": "Query Get PeersStatus",
        "description": "Query Get PeersStatus",
        "parameters": [
          {
            "name": "channelName",
            "in": "path",
            "description": "channel name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number"
                },
                "peers": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "requests": {
                        "type": "string"
                      },
                      "server_hostname": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/blockAndTxList/{channel_genesis_hash}/{number}": {
      "x-swagger-router-controller": "dbroutes",
      "get": {
        "tags": ["Transactions"],
        "operationId": "txPerBlock",
        "summary": "Query to Get Block And Transaction List",
        "description": "Query to Get Block And Transaction List",
        "parameters": [
          {
            "name": "channel_genesis_hash",
            "in": "path",
            "description": "channel_genesis_hash",
            "required": true,
            "type": "string"
          },
          {
            "name": "number",
            "in": "path",
            "description": "block number",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number"
                },
                "rows": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "blocknum": {
                        "type": "number"
                      },
                      "channel_genesis_hash": {
                        "type": "string"
                      },
                      "data_hash": {
                        "type": "string"
                      },
                      "prehash": {
                        "type": "string"
                      },
                      "channelname": {
                        "type": "string"
                      },
                      "txCount": {
                        "type": "number"
                      },
                      "createdt": {
                        "type": "string"
                      },
                      "prev_blockhash": {
                        "type": "string"
                      },
                      "blockhash": {
                        "type": "string"
                      },
                      "txhash": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/txByMinute/{channel_genesis_hash}/{hours}": {
      "x-swagger-router-controller": "dbroutes",
      "get": {
        "tags": ["Transactions"],
        "operationId": "txByMinute",
        "summary": "Query to Get Trasaction BY Minute",
        "description": "Query to Get Trasaction By Minute",
        "parameters": [
          {
            "name": "channel_genesis_hash",
            "in": "path",
            "description": "channel_genesis_hash",
            "required": true,
            "type": "string"
          },
          {
            "name": "hours",
            "in": "path",
            "description": "hours ",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number"
                },
                "rows": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/TxCount"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/txByHour/{channel_genesis_hash}/{days}": {
      "x-swagger-router-controller": "dbroutes",
      "get": {
        "tags": ["Transactions"],
        "operationId": "txByHour",
        "summary": "Query to Get Trasaction By Hour",
        "description": "Queryto  Get Trasaction BY Hour",
        "parameters": [
          {
            "name": "channel_genesis_hash",
            "in": "path",
            "description": "channel_genesis_hash",
            "required": true,
            "type": "string"
          },
          {
            "name": "days",
            "in": "path",
            "description": "days",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number"
                },
                "rows": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/TxCount"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/blocksByMinute/{channel_genesis_hash}/{hours}": {
      "x-swagger-router-controller": "dbroutes",
      "get": {
        "tags": ["Blocks"],
        "operationId": "blocksByMinute",
        "summary": "Query to Get Blocks BY Minute",
        "description": "Query to Get Blocks By Minute",
        "parameters": [
          {
            "name": "channel_genesis_hash",
            "in": "path",
            "description": "channel_genesis_hash",
            "required": true,
            "type": "string"
          },
          {
            "name": "hours",
            "in": "path",
            "description": "hours ",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number"
                },
                "rows": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/TxCount"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/blocksByHour/{channel_genesis_hash}/{days}": {
      "x-swagger-router-controller": "dbroutes",
      "get": {
        "tags": ["Blocks"],
        "operationId": "blocksByHour",
        "summary": "Query to Get Blocks By Hour",
        "description": "Query to Get Blocks By Hour",
        "parameters": [
          {
            "name": "channel_genesis_hash",
            "in": "path",
            "description": "channel_genesis_hash",
            "required": true,
            "type": "string"
          },
          {
            "name": "days",
            "in": "path",
            "description": "days",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number"
                },
                "rows": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/TxCount"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/txByOrg/{channel_genesis_hash}": {
      "x-swagger-router-controller": "dbroutes",
      "get": {
        "tags": ["Blocks"],
        "operationId": "txByOrg",
        "summary": "Query Get Trasaction By Org",
        "description": "Query Get Trasaction BY Org",
        "parameters": [
          {
            "name": "channel_genesis_hash",
            "in": "path",
            "description": "channel_genesis_hash",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number"
                },
                "rows": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "count": {
                        "type": "string"
                      },
                      "creator_msp_id": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/channels/info": {
      "x-swagger-router-controller": "dbroutes",
      "get": {
        "tags": ["Channels"],
        "operationId": "channelsInfo",
        "summary": "Query to Get Channels Inforamtion",
        "description": "Query to Get channel Information",
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number"
                },
                "channels": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "channelName": {
                        "type": "string"
                      },
                      "channel_genesis_hash": {
                        "type": "string"
                      },
                      "channel_hash": {
                        "type": "string"
                      },
                      "createdat": {
                        "type": "string"
                      },
                      "blocks": {
                        "type": "number"
                      },
                      "transactions": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/tx/getinfo/{txid}": {
      "x-swagger-router-controller": "platformroutes",
      "get": {
        "tags": ["Transactions"],
        "operationId": "transactionsInfo",
        "summary": "Query to Get Trasaction By Hour",
        "parameters": [
          {
            "name": "txid",
            "in": "path",
            "description": "txid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    }
  }
}